Fix infrastructure and build issues for Summit (#509)
* Add presets for summit

* Reset describe when rebuilding from presets

* Loosen test tolerances

Compiling with -march=native -mtune=native on Power9 gives slightly
different answers. VecGeom and ORANGE have slightly different answers,
and using -fastmath gives *very* different answers (not yet accounted
for in our test comparison so this still fails).

* Fix geo-check test when run outside of source dir

* Hide scary host code from timid NVCC

* Fix missing operator bool on cu/hip RNG

* Turn off fast math by default

* Fix attempt to build unsupported muon processes when ROOT is disabled

* Die correctly when attempting to build Coulomb process

* Use celeritas basic physics list by default

* Fix output of geant options

* Enable geo check build if vecgeom is disabled

* Fix duplicate main definitino

* Add hidden variable for forcing resource lock on tests

+++ app/geo-check/GCheckKernel.cc
@@ -9,6 +9,7 @@
 
 #include "corecel/data/CollectionStateStore.hh"
 #include "celeritas/field/LinearPropagator.hh"
+#include "celeritas/geo/GeoParams.hh"
 
 using namespace celeritas;
 
@@ -70,7 +71,7 @@ GCheckOutput run_cpu(const SPConstGeo&          params,
                     step.distance,
                     geo.find_next_step().distance,
                     physid(geo),
-                    geo.volume_id().get());
+                    static_cast<int>(geo.volume_id().get()));
             }
         }
         else
